+++ SyLife/RigidbodySearcher.h
@@ -12,7 +12,7 @@ public:
 	RigidbodySearcher();
 	~RigidbodySearcher();
 
-	std::vector<std::pair<size_t, int>>	GetNearRigidbodies(Vector2D position, double radius) const;
+	std::vector<std::pair<size_t, double>>	GetNearRigidbodies(Vector2D position, double radius) const;
 };
 
 extern unique_ptr<RigidbodySearcher> g_rigidbodySearcherPtr;
